+++ src/frontend/src/components/errors/NotAuthenticated.tsx
