+++ src/frontend/src/components/errors/NotFound.tsx
