+++ src/frontend/src/components/errors/PermissionDenied.tsx
