+++ src/frontend/src/pages/ErrorPage.tsx
@@ -3,8 +3,8 @@ import { useDocumentTitle } from '@mantine/hooks';
 import { useEffect, useState } from 'react';
 import { useRouteError } from 'react-router-dom';
 
+import GenericErrorPage from '../components/errors/GenericErrorPage';
 import { ErrorResponse } from '../states/states';
-import GenericErrorPage from './errors/GenericErrorPage';
 
 export default function ErrorPage() {
   const error = useRouteError() as ErrorResponse;
